requirements.txt
discord.py==2.3.2
yfinance==0.2.38
python-dotenv==1.0.1
matplotlib==3.8.0